Enabled more clang-tidy checks

+++ Source/NimbleCommander/NimbleCommander/States/FilePanels/Helpers/IconRepositoryCleaner.mm
@@ -16,7 +16,7 @@ void IconRepositoryCleaner::SweepUnusedSlots()
     const auto used_slots = m_Repository.AllSlots();
     auto still_in_use = std::vector<bool>(used_slots.size(), false);
 
-    for( auto i = 0, e = static_cast<int>(m_Data.RawEntriesCount()); i < e; ++i ) {
+    for( auto i = 0, e = m_Data.RawEntriesCount(); i < e; ++i ) {
         auto &vd = m_Data.VolatileDataAtRawPosition(i);
         if( vd.icon != vfsicon::IconRepository::InvalidKey ) {
             auto it = std::ranges::lower_bound(used_slots, vd.icon);
